backend/requirements.txt
# Python版本: 3.12.10
# 更新时间: 2026-04-05

# ---- Django框架 ----
Django>=6.0.2,<6.1                           # Web框架
djangorestframework>=3.15.0,<4.0           # REST API框架
djangorestframework-simplejwt>=5.4.0,<6.0  # JWT认证
django-cors-headers>=4.6.0,<5.0            # CORS跨域支持
django-filter>=25.2,<26.0                  # 过滤器支持（升级到25.2）
channels>=4.2.0,<5.0                       # WebSocket / ASGI 支持
daphne>=4.1.2,<5.0                         # ASGI 开发服务器

# ---- 数据库 ----
psycopg2-binary>=2.9.10,<3.0               # PostgreSQL驱动
neo4j>=6.1.0,<7.0                          # Neo4j图数据库驱动（升级到6.1.0）

# ---- AI/LLM ----
torch==2.11.0
scikit-learn>=1.8.0,<2.0                   # DKT训练评估（AUC等指标）
langchain>=1.2.7,<2.0                      # LangChain框架（升级到1.x）
langchain-openai>=1.1.7,<2.0               # LangChain兼容聊天客户端（用于 Qwen / DeepSeek）
langchain-community>=0.4.1,<1.0            # LangChain社区组件（升级到0.4.x）
neo4j-graphrag>=1.14.1,<2.0                # Neo4j 官方 GraphRAG Python 包
qdrant-client>=1.17.1,<2.0                 # 本地/远程 Qdrant 向量库客户端
sentence-transformers>=5.3.0,<6.0          # 可选：本地语义向量模型（GraphRAG 高质量检索）

# ---- 工具库 ----
python-dotenv>=1.0.1,<2.0                  # 环境变量管理
Pillow>=12.1.1,<13.0                       # 图像处理（升级到12.x）
openpyxl>=3.1.5,<4.0                       # Excel文件处理
pandas>=3.0.0,<4.0                         # 数据分析（升级到3.x）
requests>=2.31.0,<3.0                      # HTTP请求（tools/测试工具依赖）
xlrd>=2.0.1

# ---- API文档 ----
drf-spectacular>=0.28.0,<1.0               # OpenAPI文档生成
